+++ driver/ni/read_task.h
@@ -272,7 +272,7 @@ private:
         const auto hw_res = this->hw_reader->read(n_samples, this->buf);
         // A non-zero skew means that our application cannot keep up with the
         // hardware acquisition rate.
-        if (std::abs(hw_res.skew) > this->cfg.skew_warn_on_count)
+        if (static_cast<size_t>(std::abs(hw_res.skew)) > this->cfg.skew_warn_on_count)
             res.warning = common::skew_warning(hw_res.skew);
 
         auto prev_read_err = this->curr_read_err;
